Revert "move code into a folder hierarchy that matches the code hierarchy"
This reverts commit 64167015be1a592ce7d7705c90b5f237383caef1.
